(mini.basics) Use a more specific 'TermOpen' autocomand pattern.
The pattern that the “Start builtin terminal in Insert mode” autocommand is using is not specific enough, which seems to cause [Neogit](https://github.com/NeogitOrg/neogit) to open in insert mode which shouldn’t be possible, causing an error that stops Neogit being used until the `esc` key is hit. Changing the pattern to `term://*` stops the Neogit problem, and keeps the autocommand working.