Enhance window resize logic
Window resizing was a bit confusing, with left sometimes behaving like right and right sometimes behaving like left. This was not in line with user expectations. 

Here we enhance the logic behind, ensuring that left remains left and right remains right, similar to how it works for "nN" and "Nn."