(mini.files) Expose `go_in_plus()` and `go_out_plus()`
Expose the "plus" variants of `go_in` and `go_out` so that users may
create multiple mappings to these actions.

It is not currently possible to map more than one key to these actions
without implementing their behavior manually, since the only way to map
a key to these actions is through the `mappings` table of the `setup()`
call, which does not support mapping multiple keys to a single action.