fix(deps): update how error is detected in CLI jobs
Previously a job was declared as having an error if *either* `stderr`
was written or exit code was not 0. This proved to be a wrong approach
as `stderr` can be used to write purely diagnostic/warning messages.

For example, cloning from 'https://git.sr.ht' does not support
`--fillter=blob:none`. This results into the `filtering not recognized
by server, ignoring` being written to `stderr` all while cloning is
successful.

The more robust approach is to infer that job had an error *only* if
the exit code is not zero. In this case `stderr` should be used as
additional information about the error. Case of not trivial `stderr`
and zero exit code should be treated as CLI tool decided to print
diagnostic/warning messages.