refactor(comment): modify only necessary parts of target lines
Before it was: compute whole target lines and set them as bulk with
active `lockmarks` (to preserve regular marks).

Although this approach has significantly better performance, the
downside is that it alters extmarks. For example, any highlighting that
should be present both in regular and commented text (like in
'mini.hipatterns') disappears. Or signs added via extmarks disappear.

Now only necessary parts of lines are modified (added or deleted).

As consequence, both comment and uncomment now try to preserve as much
whitespace as possible.