test(clue): follow up 'mini.comment' refactor
Previously cursor was always put on the first column of first target
line. Now on Neovim>=0.10 it is put on just after added left comment
part of first target line.

This is due to change in how commenting is done: previously it was bulk
replace of lines, now - targeted change of only comment parts.