fix(mini-starter): use scheme name for Starter buffer
When using a session manager and in the off-chance that there's a real
souce file named 'Starter' in the session, restoring the session with
this file from mini.starter causes an issue.

This is because mini.starter buffer name conflicts with the real source
file named Starter.

This can be avoided by using a scheme-like name for the starter buffer
which is a common practice for buffers created by plugins.

Signed-off-by: ldelossa <louis.delos@gmail.com>