ci: ignore lines after the scissor line
When using `git commit --verbose` (or when the commit.verbose git config
option is set), the diff of the commit is placed into the buffer when
writing the commit message. This can cause the commit linter to fail
because of long lines. However, anything below the "scissor line" is
removed from the commit message by git (see `help git commit`) so should
not be validated by the commit linter.