ci: update commit lint to respect lines which will be cleaned up by Git
This commit add the distinction of "strict" and "not strict" contexts.
Basically, the editing commit message is not strict, while linting
already done commit messages is strict.

This is useful because some lines will be cleanud up by Git itself (see
`--cleanup` option of `git commit`). Taking them into account during
linting the commit message in Git editor does not make sense and is
harmful. The ignored lines are:
- Comments (assuming default "#" as comment symbol).
- Anything after and including "scissor" lines.
- Trailing blank lines.

NOTE: this heavily relies on the fact that `git commit` was called with
default `--cleanup` (and not, say, `--cleanup=verbatim`). Otherwise, it
will not go past CI check.

Resolve #854