fix(git): account for user's pager and forced formatting
The main source of problems is doing option completion via parsing
output of `git help --man xxx`. This not only uses help text formatted
in "man" format, but also uses user's `man` override (like through
`MANPAGER` and `PAGER` environment variables). On some systems (like
Mac) this caused option completion to hang until job timeout.

The other issue is that even with (seemingly) no MANPAGER used, the text
might still be formatted not as plain text. In particular, there can be
something like "N\bNA\bAM\bME\bE" instead of "NAME". The `x\bx` and
`_\bx` are an "old" way to format character as bold and italic. The
origin of this on Mac seems to be `groff` / `grotty` / or something like
that.

Resolve #918