feat(icons): improve coverage of `mock_nvim_web_devicons()`
I did not see the value in *actually* mocking all `get_icons_*()`
functions. Returning empty table seemed to be enough.

However, 'ibhagwan/fzf-lua' uses them and actually checks if they are
not empty. Although this might be fixed in the future (by their
explicit 'mini.icons' support), this shows that the functions might be
used. Hence, mock them properly as much as possible.