fix(move): correctly save/restore registers
`getreg()` only returns register contents, not register type, so
restoring the register contents converted the `z` register to linewise
or charwise text, depending on whether the text contains newlines. It
also converts all `NUL`s into newlines, which I'm sure was very
frustrating for people who like to put `NUL`s in their code.

The `"` was also incorrectly restored. When assigning to the `"`
register, the text actually gets saved to the `0` register and the `"`
register is pointed to the `0` register. The `"` register cannot
actually hold text, it is only a reference to one of the other
registers.

This commit fixes these issues by saving the `z` register state with
`getreginfo()` and saving and restoring only the `points_to` value of
the `"` register.

See `:h getreg()`, `:h getreginfo()`.