fix(move): avoid error when buffer has no lines
Put fails with an error (which halts execution in a Lua context) when
trying to use a register which is unset. `x` and `d` do not set their
register when they don't change the buffer, which can happen when the
buffer is empty.

This commit ensures the `z` register contains the empty string if it
might otherwise have been unset before putting from it. If the `z`
register was actually unset, this unset state gets restored at the end
of the function.