fix(files): make file manipulation work across devices
Function `vim.loop.fs_rename()` does not work across devices (i.e.
partitions in Linux and volumes in Windows) and it is an known decision.

This commit makes moving across devices work. As well as moving to trash
(which can happen if deleting file in, for example, "D:\" Windows volume
and trash is in "C:\" volume).

Resolve #1152