fix(align): correctly handle blockwise to-eol
When the blockwise selection is in to-EOL mode (when the last horizontal
motion extends to EOL, like $) the selection on each line extends to EOL
regardless of the position of the rightmost selection corner. With
'virtualedit', the selection extends to the column one character past
the end of the longest line in the region.

getcurpos() seems to be the only way to detect if the blockwise
selection is in to-EOL mode, as the last field (curswant) is v:maxcol if
the cursor is in to-EOL mode, which is normally true when the active
blockwise selection is to-EOL. Swapping to one of the left corners with
o or O deactivates to-EOL mode, so this isn't a problem. This is fine
for visual mode, but requires an awkward workaround in operator mode:
    1. Save cursor position
    2. Restore visual selection
    3. Call getcurpos()
    4. Exit visual mode
    5. Restore cursor position

There is one exception where the cursor position doesn't match the
to-EOL-ness of the selection, which is after the '> mark has been
manually set, such as via m>.

See https://github.com/vim/vim/issues/15587