test(align): fix expected screenshots
This affects ['Align with preview']['works']. The operator blockwise and
visual blockwise forms of this test should affect the text in the same
way, but the reference screenshots didn't reflect this.

The old screenshots showed the result of an erroneously exclusive
gA<C-v>`a motion. If d<C-v>`a is used instead in the same situation, the
characters after the first _ (in the same column as the 'a mark) are
deleted.

Forced-blockwise motions behave the same as visual-blockwise motions,
except the cursor is placed at the start of the affected region. This
means the motion is inclusive or exclusive depending on the 'selection'
option.