fix(align): compat for nvim < 0.10
Adds a fallback for v:maxcol for < 0.9.0. Adds workaround for getting
both left and right virtcol bounds in pos_to_virtcol() for < 0.10.0 as
the list flag for virtcol() requires 0.10.0.

Also replaces next_char_boundary() with byte_range_from_virtcols(). This
allows both call sites to be cleaner, and the new implementation is
robust against virtcol2col() returning different byte positions for
multibyte characters on different versions. On 0.10.0 and later
virtcol2col() returns the index of the first byte of the character at
the specified position, but it returns the index of the last byte on
older versions. It would actually save a lot of hassle if it were
possible to get both these values, as the weird workaround with
byteidx() and charidx() wouldn't be necessary.