fix(starter): empty buffer issue when editing recent file with swapfile
When `swapfile` is set, a file has an existing swapfile, and the file
is listed in the recent files section, editing this file and selecting
`e` or `o` in the popup swapfile menu results in an error and ultimately
leaves you with an empty buffer, maybe this problem also appears in other similar case.
Fix this by add a `silent! ` at the front of `vim.cmd` parameter in `eval_fun_or_string` function.