feat(trailspace): continue highlighting whitespace in insert mode
Continue highlighting trailing whitespace in insert mode so that
trailing whitespace is clearly visible.

To avoid highlighting spaces when typing normally, replace the standard
trailing whitespace regex with one that does negative lookbehind to
check the cursor position when entering insert mode. The regex ensures
that highlighting does not happen if cursor is at the very end of the
current line.