fix(extra): remove autocommand and preserve vim.o.background
The background needs to be preserved as previewing a theme can
inadvertently change that value if you select a light-only theme.
While previewing the next theme, if it happens to have both a
light and dark variant, then the light will be shown, even if
the user originally had 'dark' for their background.