feat(extra): incorporate mini.colors in colorscheme picker
Two options are presented. colorschemes_1 is more concise but
may result in a colorscheme being set more than once. For example,
if you preview one (sets the colorscheme) and then select that
one, the colorscheme is set again. Is this really a big deal?

colorscheme_2 is more verbose to avoid any duplicate calls to
setting the colorscheme. It seems a bit verbose just to avoid
calling colorscheme more than once.

Thoughts?